refactor(core)!: simplify get creds for proof api
BREAKING CHANGE: The `ProofsModule.getRequestedCredentialsForProofRequest` expected some low level message objects as input. This is not in line with the public API of the rest of the framework and has been simplified to only require a proof record id and optionally a boolean whether the retrieved credentials should by filtered based on the proof proposal (if avaiable).

Signed-off-by: Timo Glastra <timo@animo.id>