fix: allow agent without inbound endpoint to connect when using multi-use invitation (#712)
Previously the session is only stored by the connection id, but this is not always enough for the session to be reused. When using a multi-use invitation the connection id will change while processing the message, meaning the session cannot be reused. This also helps with cases where no connection is established. Fixes #483

Signed-off-by: Timo Glastra <timo@animo.id>