feat!: use did:key in protocols by default (#1149)
Signed-off-by: Ariel Gentile <gentilester@gmail.com>

BREAKING CHANGE: `useDidKeyInProtocols` configuration parameter is now enabled by default. If your agent only interacts with modern agents (e.g. AFJ 0.2.5 and newer) this will not represent any issue. Otherwise it is safer to explicitly set it to `false`. However, keep in mind that we expect this setting to be deprecated in the future, so we encourage you to update all your agents to use did:key.